Move vote program state and instructions to solana-program
