+++ programs/vote/Cargo.toml
@@ -19,6 +19,7 @@ serde_derive = "1.0.103"
 solana-frozen-abi = { path = "../../frozen-abi", version = "=1.12.0" }
 solana-frozen-abi-macro = { path = "../../frozen-abi/macro", version = "=1.12.0" }
 solana-metrics = { path = "../../metrics", version = "=1.12.0" }
+solana-program = { path = "../../sdk/program", version = "=1.12.0" }
 solana-program-runtime = { path = "../../program-runtime", version = "=1.12.0" }
 solana-sdk = { path = "../../sdk", version = "=1.12.0" }
 thiserror = "1.0"
