+++ programs/vote/src/lib.rs
@@ -1,9 +1,6 @@
 #![cfg_attr(RUSTC_WITH_SPECIALIZATION, feature(min_specialization))]
 #![allow(clippy::integer_arithmetic)]
 
-pub mod authorized_voters;
-pub mod vote_error;
-pub mod vote_instruction;
 pub mod vote_processor;
 pub mod vote_state;
 pub mod vote_transaction;
@@ -14,4 +11,7 @@ extern crate solana_metrics;
 #[macro_use]
 extern crate solana_frozen_abi_macro;
 
-pub use solana_sdk::vote::program::{check_id, id};
+pub use solana_sdk::vote::{
+    authorized_voters, error as vote_error, instruction as vote_instruction,
+    program::{check_id, id},
+};
