+++ runtime/src/bank.rs
@@ -10186,13 +10186,13 @@ pub(crate) mod tests {
         bank0.store_account_and_update_capitalization(&stake_id, &stake_account);
 
         // generate some rewards
-        let mut vote_state = Some(VoteState::from(&vote_account).unwrap());
+        let mut vote_state = Some(vote_state::from(&vote_account).unwrap());
         for i in 0..MAX_LOCKOUT_HISTORY + 42 {
             if let Some(v) = vote_state.as_mut() {
-                v.process_slot_vote_unchecked(i as u64)
+                vote_state::process_slot_vote_unchecked(v, i as u64)
             }
             let versioned = VoteStateVersions::Current(Box::new(vote_state.take().unwrap()));
-            VoteState::to(&versioned, &mut vote_account).unwrap();
+            vote_state::to(&versioned, &mut vote_account).unwrap();
             bank0.store_account_and_update_capitalization(&vote_id, &vote_account);
             match versioned {
                 VoteStateVersions::Current(v) => {
@@ -10307,13 +10307,13 @@ pub(crate) mod tests {
         bank.store_account_and_update_capitalization(&stake_id2, &stake_account2);
 
         // generate some rewards
-        let mut vote_state = Some(VoteState::from(&vote_account).unwrap());
+        let mut vote_state = Some(vote_state::from(&vote_account).unwrap());
         for i in 0..MAX_LOCKOUT_HISTORY + 42 {
             if let Some(v) = vote_state.as_mut() {
-                v.process_slot_vote_unchecked(i as u64)
+                vote_state::process_slot_vote_unchecked(v, i as u64)
             }
             let versioned = VoteStateVersions::Current(Box::new(vote_state.take().unwrap()));
-            VoteState::to(&versioned, &mut vote_account).unwrap();
+            vote_state::to(&versioned, &mut vote_account).unwrap();
             bank.store_account_and_update_capitalization(&vote_id, &vote_account);
             match versioned {
                 VoteStateVersions::Current(v) => {
@@ -15726,10 +15726,10 @@ pub(crate) mod tests {
         vote_pubkey: &Pubkey,
     ) {
         let mut vote_account = bank.get_account(vote_pubkey).unwrap_or_default();
-        let mut vote_state = VoteState::from(&vote_account).unwrap_or_default();
+        let mut vote_state = vote_state::from(&vote_account).unwrap_or_default();
         vote_state.last_timestamp = timestamp;
         let versioned = VoteStateVersions::new_current(vote_state);
-        VoteState::to(&versioned, &mut vote_account).unwrap();
+        vote_state::to(&versioned, &mut vote_account).unwrap();
         bank.store_account(vote_pubkey, &vote_account);
     }
