+++ sdk/program/src/vote/state/vote_state_0_23_5.rs
@@ -1,3 +1,4 @@
+#![allow(clippy::integer_arithmetic)]
 use super::*;
 
 const MAX_ITEMS: usize = 32;
